EDA_Master/requirements.txt
# Core Dependencies
PyQt6>=6.4.0
pandas>=1.5.0
numpy>=1.23.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# AI Integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=0.21.0

# Optional: Data Profiling
# ydata-profiling>=4.0.0

# Optional: Advanced ML
# catboost>=1.1.0
# optuna>=3.0.0
